+++ dateparser/date_parser.py
@@ -31,7 +31,11 @@ class new_parser(parser.parser):
             timestr = timestr.decode()
 
         # Parse timestr
-        res = self._parse(timestr, **kwargs)
+        # handle dateutil>=2.5 tuple result first
+        try:
+            res, _ = self._parse(timestr, **kwargs)
+        except TypeError:
+            res = self._parse(timestr, **kwargs)
 
         if res is None:
             raise ValueError("unknown string format")
@@ -182,7 +186,8 @@ class DateParser(object):
         if tz is not None:
             date_obj = tz.localize(date_obj)
 
-        date_obj = apply_timezone(date_obj, settings.TIMEZONE)
+        if settings.TIMEZONE:
+            date_obj = apply_timezone(date_obj, settings.TIMEZONE)
 
         if not settings.RETURN_AS_TIMEZONE_AWARE:
             date_obj = date_obj.replace(tzinfo=None)