+++ dateparser/freshness_date_parser.py
@@ -17,14 +17,20 @@ PATTERN = re.compile(r'(\d+)\s*(%s)\b' % _UNITS, re.I | re.S | re.U)
 
 class FreshnessDateDataParser(object):
     """ Parses date string like "1 year, 2 months ago" and "3 hours, 50 minutes ago" """
+    def __init__(self):
+        self._now = None
 
     @property
     def now(self):
-        return datetime.utcnow()
+        return self._now if self._now else datetime.utcnow()
+
+    @now.setter
+    def now(self, value):
+        self._now = value
 
     def _are_all_words_units(self, date_string):
         skip = [_UNITS,
-                r'ago|\d+',
+                r'ago|in|\d+',
                 r':|[ap]m']
 
         date_string = re.sub(r'\s+', ' ', date_string.strip())
@@ -36,7 +42,7 @@ class FreshnessDateDataParser(object):
     def _parse_time(self, date_string):
         """Attemps to parse time part of date strings like '1 day ago, 2 PM' """
         date_string = PATTERN.sub('', date_string)
-        date_string = re.sub(r'\bago\b', '', date_string)
+        date_string = re.sub(r'\b(?:ago|in)\b', '', date_string)
         if is_dateutil_result_obj_parsed(date_string):
             try:
                 return parse(date_string).time()
@@ -44,6 +50,9 @@ class FreshnessDateDataParser(object):
                 pass
 
     def parse(self, date_string, settings):
+        if settings.RELATIVE_BASE:
+            self.now = settings.RELATIVE_BASE
+
         date, period = self._parse(date_string)
 
         if date:
@@ -54,11 +63,12 @@ class FreshnessDateDataParser(object):
             else:
                 # No timezone shift takes place if time is given in the string.
                 # e.g. `2 days ago at 1 PM`
-                date = apply_timezone(date, settings.TIMEZONE)
+                if settings.TIMEZONE:
+                    date = apply_timezone(date, settings.TIMEZONE)
 
             if not settings.RETURN_AS_TIMEZONE_AWARE:
                 date = date.replace(tzinfo=None)
-
+        self.now = None
         return date, period
 
     def _parse(self, date_string):
@@ -77,7 +87,10 @@ class FreshnessDateDataParser(object):
                     break
 
         td = relativedelta(**kwargs)
-        date = self.now - td
+        if re.search(r'\bin\b', date_string):
+            date = self.now + td
+        else:
+            date = self.now - td
         return date, period
 
     def get_kwargs(self, date_string):